requirements.txt
aiosmtplib==2.0.1
anyio==3.6.2
bcrypt==4.0.1
blinker==1.5
cffi==1.15.1
click==8.1.3
cryptography==39.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.93.0
fastapi-mail==1.2.6
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
motor==3.1.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
pymongo==4.3.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
